# Improve performance of SELECTs with active mutations