@azat 2024-02-11
>AST fuzzer (asan) — Logical error: 'Cannot capture column 2 because it has incompatible type: got UInt8, but Nullable(UInt8) is expected.'.

- https://github.com/ClickHouse/ClickHouse/issues/59809

>[Stateless tests (debug) [2/5] — fail: 3, passed: 1174, skipped: 10](https://s3.amazonaws.com/clickhouse-test-reports/59531/1253a73a4556c86d98c782c5764e5aa8acddd748/stateless_tests__debug__[2_5].html)

Timeouts:

- 02782_uniq_exact_parallel_merging_bug
- 01459_manual_write_to_replicas
- 01171_mv_select_insert_isolation_long