@al13n321 2024-02-16
Oops, turns out the cloud version of CH actually implements other mutation types in AlterConversions, not just RENAME_COLUMN (https://clickhouse.com/docs/en/guides/developer/lightweight-update).

I pushed a commit to this PR that adds `AlterConversions::supportsMutationCommandType()` and uses it instead of directly checking for RENAME_COLUMN in `getAlterMutationCommandsForPart()`, to make it less error-prone to maintain. And also flattened `vector<MutationCommands>` to just `MutationCommands` along the way. Feel free to review the changes, or revert them and push something else, or whatever.