@azat 2024-02-16
LGTM, though clang-tidy fails:

>Feb 16 03:10:25 /build/src/Storages/MergeTree/ReplicatedMergeTreeQueue.cpp:1814:41: error: Dereference of undefined pointer value [clang-analyzer-core.NullDereference,-warnings-as-errors]
>Feb 16 03:10:25  1814 |             for (const auto & command : mutation_status->entry->commands | std::views::reverse)