@al13n321 2024-02-19
Huh, that warning makes absolutely no sense to me. Shuffled the code a little to work around it, guess it's a clang issue.